% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{compartment_defn}
\alias{compartment_defn}
\title{Compartment definitions}
\format{
A data.frame of definitions
}
\usage{
compartment_defn
}
\description{
A data.frame with the details around the different compartment
used in the deterministic models.
}
\keyword{datasets}
