% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/detect_infections_ode.R
\name{detect_infections_ode}
\alias{detect_infections_ode}
\title{Detect Infection from ODE Solution}
\usage{
detect_infections_ode(
  model,
  prop_symptomatic = NULL,
  prop_production_drop = NULL,
  n_detected = NULL,
  n_infected = NULL,
  ...
)
}
\arguments{
\item{model}{a solution from an ODE}

\item{prop_symptomatic}{a numeric indicating the proportion threshold
for detection considering only those in the "B" class or symptomatic.}

\item{prop_production_drop}{a numeric indicating the proportion drop in
production to indicate an outbreak}

\item{n_detected}{a numeric indicating the number of detected infections
which considers all of those invididuals existing in the R, Ra, A, I, and B
classes.}

\item{n_infected}{a numeric indicating the number of actively infected
individuals before detection occurs.}
}
\value{
a list object with the time points associated with the detection
approach
}
\description{
The purpose of this function is to detect an outbreak
under different scenarios and uses of detection.
}
