% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_det_ode.R
\name{format_det_ode_output}
\alias{format_det_ode_output}
\title{Format deterministic output}
\usage{
format_det_ode_output(dat, states = compartment_defn$state)
}
\arguments{
\item{dat}{a dataset generate from the function}

\item{states}{a vector representing the states used in the compartmental model}
}
\value{
a data.frame formatted with the appropriate column names
}
\description{
Format deterministic output
}
