% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_outbreak_chart.R
\name{generate_outbreak_chart}
\alias{generate_outbreak_chart}
\title{Generate the outbreak chart}
\usage{
generate_outbreak_chart(
  data = here::here("data-raw", "2025-11-21-h5n1-dairy.csv")
)
}
\arguments{
\item{data}{The path to the data file.}
}
\value{
A ggplot object.
}
\description{
This function generates the outbreak chart from the data-raw/2025-11-21-h5n1-dairy.csv file.
}
