% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/h5n1speed-package.R
\docType{package}
\name{h5n1speed-package}
\alias{h5n1speed}
\alias{h5n1speed-package}
\title{h5n1speed: Understanding the role of speed on containing outbreaks on dairy farms.}
\description{
What the package does (one paragraph).
}
\seealso{
Useful links:
\itemize{
  \item \url{https://wf-id.github.io/h5n1speed/}
}

}
\author{
\strong{Maintainer}: Michael DeWitt \email{medewitt@wakehealth.edu} (\href{https://orcid.org/0000-0001-8940-1967}{ORCID})

}
\keyword{internal}
