% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_asym_r0.R
\name{make_asym_r0}
\alias{make_asym_r0}
\title{Create R0 for asymptomatic infections}
\usage{
make_asym_r0(p, ffi)
}
\arguments{
\item{p}{a number representing the proprortion of asymptomatic infections}

\item{ffi}{a number, final fraction infected}
}
\description{
Create R0 for asymptomatic infections
}
\examples{
make_asym_r0(0.1, .4)
stopifnot(make_asym_r0(0.0, .4) - 1 / (1 - .4) < 1e-10)

plot(seq(0, .6, 0.01), 
  sapply(seq(0, .6, 0.01), 
  function(x) make_asym_r0(x, 0.4)), 
  ylim = c(1, 10), 
  type = "l", xlab = "Proportional asymptomatic", ylab = expression(R[0]))

}
