% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_det_ode.R
\name{plot.rRsurveillace_deterministic}
\alias{plot.rRsurveillace_deterministic}
\title{Ploting function}
\usage{
\method{plot}{rRsurveillace_deterministic}(x, ...)
}
\arguments{
\item{x}{an object}

\item{...}{other parameters}
}
\description{
Ploting function
}
