% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/R0_to_r.R
\name{r_to_gamma}
\alias{r_to_gamma}
\title{Simple conversion of R0 to r}
\usage{
r_to_gamma(r, r_0 = 2.73)
}
\arguments{
\item{r}{a number representing the instrinic growth rate}

\item{r_0}{a number, the target basic reproduction number}
}
\description{
Given that r = 1/ phi ($R_0$ - 1) then we can solve for phi
given some value of $R_0$ by phi = (R_0 -1) / r.
}
\examples{
r_to_gamma(1, 2)
r_to_gamma(1, 4)

# Estimate COVID-19
r_to_gamma(c(.22, 0.2, 0.24), 2.5)

# H1N1 2009
r_to_gamma(0.37, 2.3)

}
