% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_det_ode.R
\name{run_det_ode}
\alias{run_det_ode}
\title{Run ODE}
\usage{
run_det_ode(
  init.inf = 1,
  herd.size = 500,
  r0 = 1.2,
  fever.duration = 1.15,
  milk.loss.dur = 7,
  cow.prod.lifespan = 365.25 * 2,
  include.prod.lifespan = FALSE,
  base.milk.prod = 100,
  sympt.milk.prod = 100 - 25,
  recovered.milk.prod = 80,
  discard.sick.prod = FALSE,
  disease.induced.mortality = 0,
  prop.cull = 0,
  sim.length = 365,
  gamma_multiplier = 0,
  beta_multiplier = 0,
  intervention_date = NULL,
  intervention_stop = sim.length,
  p_asymptomatic = 0,
  sigmoid_fun = FALSE,
  add_milk = TRUE,
  transmission.type = "density"
)
}
\arguments{
\item{init.inf}{integer, the number of initial infection}

\item{herd.size}{integer, the size of the herd}

\item{r0}{positive number, the basic reproduction number}

\item{fever.duration}{a positive number, the days of fever (infectious)}

\item{milk.loss.dur}{a positive number, the days of milk loss}

\item{cow.prod.lifespan}{a positive number, the productive lifespan of a cow
in days.}

\item{include.prod.lifespan}{a logical, should background demographics
be considered with a default of \code{FALSE}}

\item{base.milk.prod}{a positive number, the baseline milk production}

\item{sympt.milk.prod}{a positive number, the milk production when the cow
is symptomatic}

\item{recovered.milk.prod}{a positive number, the production among recovered
cows.}

\item{discard.sick.prod}{a logical, discard infected milk from production}

\item{disease.induced.mortality}{a positive number between 0 and 1
representing the mortality rate from disease.}

\item{prop.cull}{a positive number between 0 and 1 representing the
proportion of cows who recover from infection who are sent to market}

\item{sim.length}{a positive number, number of days to run the simulation}

\item{gamma_multiplier}{a number, representing the impact on the recovery
rate where values greater than 0 represent faster removal and values
less than 0 represent longer recovery times.}

\item{beta_multiplier}{a number, represent the impact on the contact rate
rate where values less than 0 indicate less contact while values
greater than 0 indicate more contact.}

\item{intervention_date}{a positive number represent the day number of an
an intervention with a default of \code{NULL} representing no
intervention taking place.}

\item{intervention_stop}{a positive number representing the day that the
intervention in the beta_multiple stops. Removal rates are assumed to
continue after the intervention and stop dates.}

\item{p_asymptomatic}{a number between 0 and 1, the proportion of the transmission
that is asymptomatic.}

\item{sigmoid_fun}{a logical, if true a sigmoid function will be utilized to
phase in the gamma mulipltier based on the intervention date.}

\item{add_milk}{a logical, if true milk production will be added to the
simulation.}

\item{transmission.type}{a string, one of "frequency" for frequency-dependent
or "density" for density-dependent transmission modality. Default of
"density."}
}
\description{
Run ordinary differential equation model
}
