% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_odin_vaccine.R
\name{run_odin_vaccines}
\alias{run_odin_vaccines}
\title{Run odin vaccine}
\usage{
run_odin_vaccines(
  intervention_time = 10,
  n_sims = 1,
  r0 = 1.2,
  gamma_in = 1/1.15,
  kappa_in = 0.17,
  zeta_in = 0.1,
  duration = 50,
  s_initial = 500,
  i_initial = 1,
  inter_scenario = 1,
  n_detected = 5
)
}
\arguments{
\item{intervention_time}{an interger number representing when
the intervention occurs}

\item{n_sims}{an integer number of simulation}

\item{r0}{a number representing the basic reproduction number}

\item{gamma_in}{a number the recovery rate and period of infectiousness}

\item{kappa_in}{a number, the rate of transition from I to B}

\item{zeta_in}{a number, the vaccination rate from S to R}

\item{duration}{an integer representing the duration of the simulation}

\item{s_initial}{a number, the initial population size}

\item{i_initial}{a number, the initial number infected}

\item{inter_scenario}{different scenarios}

\item{n_detected}{an interger, if using the second intervention the threshold}
}
\description{
Run the odin model for vaccination representing flows from
S -> I -> B -> R.
}
