% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_det_ode.R
\name{sigmoid_fun}
\alias{sigmoid_fun}
\title{Sigmoid function}
\usage{
sigmoid_fun(x, x0 = 4, k = 2)
}
\arguments{
\item{x}{a double, the timepoint at which to evaluate the function}

\item{x0}{a double, the starting timepoint of the sigmoid}

\item{k}{a double, representing the speed of the sigmoid to 1}
}
\value{
a double
}
\description{
A sigmoid function
}
