% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/h5n1speed-package.R
\docType{package}
\name{h5n1speed-package}
\alias{h5n1speed}
\alias{h5n1speed-package}
\title{h5n1speed: Understanding the role of speed and asymptomatic transmission on containing outbreaks on dairy farms}
\description{
Functions and code used to support the analysis in "Time is of the essence: effectiveness of dairy farm control for H5N1 is limited by fast spread."
}
\author{
\strong{Maintainer}: Anonymous Author \email{anonymous@example.com}

}
\keyword{internal}
