% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_intervention_ode.R
\name{run_intervention_ode}
\alias{run_intervention_ode}
\title{Estimate ODE Intervention outcomes}
\usage{
run_intervention_ode(
  r0_in = 1.2,
  gamma_in = 1/1.15,
  gamma_post = 1/(12/24),
  gamma_post_asymptomatic = gamma_in,
  sim_duration = 120,
  kappa = 1/7,
  delay_time = 7,
  p_asymptomatic = 0,
  s_ini = 500 - 1,
  i_ini = 1,
  r_ini = 0,
  phi = 0,
  ...
)
}
\arguments{
\item{r0_in}{R0 input}

\item{gamma_in}{pre-intervention gamma}

\item{gamma_post}{post-intervention gamma}

\item{sim_duration}{how long to run simulation}

\item{kappa}{the symptomatc duration}

\item{delay_time}{the time to delay an intervention}

\item{s_ini}{the initial size}

\item{i_ini}{the initial infected}

\item{r_ini}{the initial recovered}

\item{phi}{the proportion of the population that dies from the disease}

\item{...}{
  Arguments passed on to \code{\link[=detect_infections_ode]{detect_infections_ode}}
  \describe{
    \item{\code{model}}{a solution from an ODE}
    \item{\code{prop_symptomatic}}{a numeric indicating the proportion threshold
for detection considering only those in the "B" class or symptomatic.}
    \item{\code{prop_production_drop}}{a numeric indicating the proportion drop in
production to indicate an outbreak}
    \item{\code{n_detected}}{a numeric indicating the number of detected infections
which considers all of those invididuals existing in the R, Ra, A, I, and B
classes.}
    \item{\code{n_infected}}{a numeric indicating the number of actively infected
individuals before detection occurs.}
  }}
}
\value{
a tibble containing the outputs from the deSolve package of the ode include the intervention.
}
\description{
Simulating the outcomes of the ODE system when an intervention of
some intensity gamma_post is applied at some time point given the threshold
of detection and applied at some delay from identification.
}
